졸음 코딩

+++ src/main/java/com/ggonandkids/zzol/domain/challenge/presentation/dto/response/ChallengeResponseDto.java
@@ -10,7 +10,7 @@ public class ChallengeResponseDto {
     private Long challengeId;
     private String title;
     private String body;
-    private int time;
+    private float time;
     private int attend;
     private String id;
